ask librarian function is updated: student email is added in topic field

+++ src/main/java/com/iutlibrary/backend/bookStuff/bookItem/BookItemRepository.java
@@ -25,5 +25,5 @@ public interface BookItemRepository extends JpaRepository<BookItem, Long> {
     List<BookItem> findByISBN(Long isbn);
 
     @Query("UPDATE BookItem b SET b.status = ?2 WHERE b.barcode = ?1")
-    int updateStatus(Long barcode, BookStatus newStatus);
+    long updateStatus(Long barcode, BookStatus newStatus);
 }
